requirements.txt
# Core dependencies
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
google-generativeai>=0.8.0,<1.0.0
html5lib>=1.1,<2.0
py-ocsf-models>=1.0.0,<2.0.0

# Development dependencies
pytest>=8.0.0,<9.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.12.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
